Refactor process management to separate namespace
